// src/app/core/enums/codigoTipoVeiculo.ts
export enum CodigoTipoVeiculoEnum {
    CICLOMOTOR = "2",
    MOTONETA = "3",
    MOTOCICLETA = "4",
    TRICICLO = "5",
    AUTOMOVEL = "6",
    MICRO_ONIBUS = "7",
    ONIBUS = "8",
    REBOQUE = "10",
    SEMI_REBOQUE = "11",
    CAMIONETA = "13",
    CAMINHAO = "14",
    CAMINHAO_TRATOR = "17",
    TRATOR_DE_RODAS = "18",
    TRATOR_DE_ESTEIRAS = "19",
    TRATOR_MISTO = "20",
    QUADRICICLO = "21",
    CHASSI_PLATAFORMA = "22",
    CAMINHONETE = "23",
    UTILITARIO = "25",
    MOTOR_HOME = "26",
    SEM_TIPO = "100"
}

export const CodigoTipoVeiculoDescricao = [
    { value: CodigoTipoVeiculoEnum.CICLOMOTOR, label: 'Ciclomotor' },
    { value: CodigoTipoVeiculoEnum.MOTONETA, label: 'Motoneta' },
    { value: CodigoTipoVeiculoEnum.MOTOCICLETA, label: 'Motocicleta' },
    { value: CodigoTipoVeiculoEnum.TRICICLO, label: 'Triciclo' },
    { value: CodigoTipoVeiculoEnum.AUTOMOVEL, label: 'Automóvel' },
    { value: CodigoTipoVeiculoEnum.MICRO_ONIBUS, label: 'Micro Ônibus' },
    { value: CodigoTipoVeiculoEnum.ONIBUS, label: 'Ônibus' },
    { value: CodigoTipoVeiculoEnum.REBOQUE, label: 'Reboque' },
    { value: CodigoTipoVeiculoEnum.SEMI_REBOQUE, label: 'Semi Reboque' },
    { value: CodigoTipoVeiculoEnum.CAMIONETA, label: 'Camioneta' },
    { value: CodigoTipoVeiculoEnum.CAMINHAO, label: 'Caminhão' },
    { value: CodigoTipoVeiculoEnum.CAMINHAO_TRATOR, label: 'Caminhão Trator' },
    { value: CodigoTipoVeiculoEnum.TRATOR_DE_RODAS, label: 'Trator de Rodas' },
    { value: CodigoTipoVeiculoEnum.TRATOR_DE_ESTEIRAS, label: 'Trator de Esteiras' },
    { value: CodigoTipoVeiculoEnum.TRATOR_MISTO, label: 'Trator Misto' },
    { value: CodigoTipoVeiculoEnum.QUADRICICLO, label: 'Quadriciclo' },
    { value: CodigoTipoVeiculoEnum.CHASSI_PLATAFORMA, label: 'Chassi Plataforma' },
    { value: CodigoTipoVeiculoEnum.CAMINHONETE, label: 'Caminhonete' },
    { value: CodigoTipoVeiculoEnum.UTILITARIO, label: 'Utilitário' },
    { value: CodigoTipoVeiculoEnum.MOTOR_HOME, label: 'Motor Home' },
    { value: CodigoTipoVeiculoEnum.SEM_TIPO, label: 'Sem Tipo' }
];